+++ test/shallowbackup.js
@@ -20,8 +20,15 @@ const backup = require('../includes/shallowbackup.js');
 const fs = require('fs');
 const nock = require('nock');
 
+// Note all these tests include a body parameter of include_docs and a query
+// string of include_docs because of a quirk of nano that when using the fetch
+// method always adds the include_docs query string.
 describe('#unit Perform backup using shallow backup', function() {
   const dbUrl = 'http://localhost:5984/animaldb';
+  // Query string keys are stringified by Nano
+  const badgerKey = JSON.stringify('badger\0');
+  const kookaburraKey = JSON.stringify('kookaburra\0');
+  const snipeKey = JSON.stringify('snipe\0');
 
   beforeEach('Reset nocks', function() {
     nock.cleanAll();
@@ -30,20 +37,20 @@ describe('#unit Perform backup using shallow backup', function() {
   it('should perform a shallow backup', function(done) {
     var couch = nock(dbUrl)
       // batch 1
-      .get('/_all_docs')
+      .post('/_all_docs')
       .query({limit: 3, include_docs: true})
       .reply(200, JSON.parse(fs.readFileSync('./test/fixtures/animaldb_all_docs_1.json', 'utf8')))
       // batch 2
-      .get('/_all_docs')
-      .query({limit: 3, include_docs: true, startkey: '"badger\\u0000"'})
+      .post('/_all_docs')
+      .query({limit: 3, startkey: badgerKey, include_docs: true})
       .reply(200, JSON.parse(fs.readFileSync('./test/fixtures/animaldb_all_docs_2.json', 'utf8')))
       // batch 3
-      .get('/_all_docs')
-      .query({limit: 3, include_docs: true, startkey: '"kookaburra\\u0000"'})
+      .post('/_all_docs')
+      .query({limit: 3, startkey: kookaburraKey, include_docs: true})
       .reply(200, JSON.parse(fs.readFileSync('./test/fixtures/animaldb_all_docs_3.json', 'utf8')))
       // batch 4
-      .get('/_all_docs')
-      .query({limit: 3, include_docs: true, startkey: '"snipe\\u0000"'})
+      .post('/_all_docs')
+      .query({limit: 3, startkey: snipeKey, include_docs: true})
       .reply(200, JSON.parse(fs.readFileSync('./test/fixtures/animaldb_all_docs_4.json', 'utf8')));
 
     backup(dbUrl, 3, 1, null, null)
@@ -67,24 +74,24 @@ describe('#unit Perform backup using shallow backup', function() {
   it('should perform a shallow backup with transient error', function(done) {
     var couch = nock(dbUrl)
       // batch 1
-      .get('/_all_docs')
+      .post('/_all_docs')
       .query({limit: 3, include_docs: true})
       .reply(200, JSON.parse(fs.readFileSync('./test/fixtures/animaldb_all_docs_1.json', 'utf8')))
       // batch 2
-      .get('/_all_docs')
-      .query({limit: 3, include_docs: true, startkey: '"badger\\u0000"'})
+      .post('/_all_docs')
+      .query({limit: 3, startkey: badgerKey, include_docs: true})
       .reply(200, JSON.parse(fs.readFileSync('./test/fixtures/animaldb_all_docs_2.json', 'utf8')))
       // batch 3 - transient error
-      .get('/_all_docs')
-      .query({limit: 3, include_docs: true, startkey: '"kookaburra\\u0000"'})
+      .post('/_all_docs')
+      .query({limit: 3, startkey: kookaburraKey, include_docs: true})
       .reply(500, {error: 'Internal Server Error'})
       // batch 3 - retry
-      .get('/_all_docs')
-      .query({limit: 3, include_docs: true, startkey: '"kookaburra\\u0000"'})
+      .post('/_all_docs')
+      .query({limit: 3, startkey: kookaburraKey, include_docs: true})
       .reply(200, JSON.parse(fs.readFileSync('./test/fixtures/animaldb_all_docs_3.json', 'utf8')))
       // batch 4
-      .get('/_all_docs')
-      .query({limit: 3, include_docs: true, startkey: '"snipe\\u0000"'})
+      .post('/_all_docs')
+      .query({limit: 3, startkey: snipeKey, include_docs: true})
       .reply(200, JSON.parse(fs.readFileSync('./test/fixtures/animaldb_all_docs_4.json', 'utf8')));
 
     var errCount = 0;
@@ -112,16 +119,13 @@ describe('#unit Perform backup using shallow backup', function() {
   it('should fail to perform a shallow backup on fatal error', function(done) {
     var couch = nock(dbUrl)
       // batch 1
-      .get('/_all_docs')
-      .query({limit: 3, include_docs: true})
+      .post('/_all_docs').query({limit: 3, include_docs: true})
       .reply(200, JSON.parse(fs.readFileSync('./test/fixtures/animaldb_all_docs_1.json', 'utf8')))
       // batch 2
-      .get('/_all_docs')
-      .query({limit: 3, include_docs: true, startkey: '"badger\\u0000"'})
+      .post('/_all_docs').query({limit: 3, startkey: badgerKey, include_docs: true})
       .reply(200, JSON.parse(fs.readFileSync('./test/fixtures/animaldb_all_docs_2.json', 'utf8')))
       // batch 3 - fatal error
-      .get('/_all_docs')
-      .query({limit: 3, include_docs: true, startkey: '"kookaburra\\u0000"'})
+      .post('/_all_docs').query({limit: 3, startkey: kookaburraKey, include_docs: true})
       .reply(401, {error: 'Unauthorized'});
 
     var errCount = 0;